Rename Humanity component to Withering and delete Status component imports in Show.vue
