+++ resources/js/Pages/Character/Partials/Status/Withering.vue
